+++ loadtest/config.json
@@ -1,7 +1,7 @@
 {
-    "batch_size": 10,
+    "msgs_per_tx": 10,
     "chain_id": "sei-loadtest-testnet",
-    "orders_per_block": 400,
+    "txs_per_block": 6000,
     "rounds": 5,
     "price_distribution": {
         "min": "45",
