+++ loadtest/sign.go
@@ -51,7 +51,7 @@ func GetKey(accountIdx uint64) cryptotypes.PrivKey {
 	return algo.Generate()(derivedPriv)
 }
 
-func SignTx(txBuilder *client.TxBuilder, privKey cryptotypes.PrivKey, seqDelta uint64) {
+func SignTx(ChainID string, txBuilder *client.TxBuilder, privKey cryptotypes.PrivKey, seqDelta uint64) {
 	var sigsV2 []signing.SignatureV2
 	accountNum, seqNum := GetAccountNumberSequenceNumber(privKey)
 	seqNum += seqDelta
